+++ custom/Gaffer/Gaffer.py
@@ -1,3 +1,4 @@
+#!/usr/bin/env python3
 ##########################################################################
 #
 #  Copyright (c) 2019, Hypothetical Inc. All rights reserved.
@@ -126,10 +127,10 @@ class GafferPlugin(DeadlinePlugin):
         tempSceneDirectory = self.CreateTempDirectory("thread" + str(self.GetThreadNumber()))
         tempSceneFilename = Path.Combine(tempSceneDirectory, Path.GetFileName(localScript))
 
-        with open(localScript, "r") as inFile, open(tempSceneFilename, "w") as outFile:
+        with open(localScript, "r", encoding="utf-8") as inFile, open(tempSceneFilename, "w", encoding="utf-8") as outFile:
             for line in inFile:
-                newLine = RepositoryUtils.CheckPathMapping(line.decode("utf-8"))
-                outFile.write(newLine.encode("utf-8"))
+                newLine = RepositoryUtils.CheckPathMapping(line)
+                outFile.write(newLine)
         
         self._gafferScript = tempSceneFilename